serverとuserの区別をなくす
