--- user/config.ini
@@ -1,2 +1,0 @@
-[DEFAULT]
-port=5000
